+++ bindings/pydeck-carto/pydeck_carto/__init__.py
@@ -1,5 +1,10 @@
 from ._version import __version__
-from .layer import register_carto_layer
-from .carto_auth.auth import CartoAuth, CredentialsError
+from .layer import register_carto_layer, get_layer_credentials
+from .error import get_error_notifier
 
-__all__ = ["__version__", "register_carto_layer", "CartoAuth", "CredentialsError"]
+__all__ = [
+    "__version__",
+    "register_carto_layer",
+    "get_layer_credentials",
+    "get_error_notifier",
+]