+++ modules/carto/src/layers/schema/spatialjson-utils.ts
@@ -1,4 +1,4 @@
-import {bigIntToIndex} from '../quadbin-utils';
+import {bigIntToHex} from 'quadbin';
 
 export type IndexScheme = 'h3' | 'quadbin';
 type TypedArray = Float32Array | Float64Array;
@@ -13,16 +13,16 @@ export type Cells = {
 };
 export type SpatialBinary = {scheme?: IndexScheme; cells: Cells};
 export type SpatialJson = {
-  id: string;
+  id: string | bigint;
   properties: Properties;
 }[];
 
 export function binaryToSpatialjson(binary: SpatialBinary): SpatialJson {
-  const {cells} = binary;
+  const {cells, scheme} = binary;
   const count = cells.indices.value.length;
   const spatial: any[] = [];
   for (let i = 0; i < count; i++) {
-    const id = bigIntToIndex(cells.indices.value[i]);
+    const id = scheme === 'h3' ? bigIntToHex(cells.indices.value[i]) : cells.indices.value[i];
 
     const properties = {...cells.properties[i]};
     for (const key of Object.keys(cells.numericProps)) {